backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pillow==9.5.0
numpy<2.0
opencv-python==4.7.0.72
transformers==4.28.1
torch==2.0.0
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7